Add base widget class
