+++ includes/widgets/class-product.php
@@ -10,7 +10,14 @@ namespace FastWC\Widgets;
 /**
  * Fast PDP button widget class.
  */
-class Product extends \WP_Widget {
+class Product extends Widget {
+
+	/**
+	 * Widget template name.
+	 *
+	 * @var string
+	 */
+	protected $template = 'widgets/fast-product';
 
 	/**
 	 * Widget constructor.
@@ -26,24 +33,13 @@ class Product extends \WP_Widget {
 	}
 
 	/**
-	 * Process the widget options and display the HTML on the page.
+	 * Function to determine if this widget should be hidden.
 	 *
-	 * @param array $args     Widget args.
 	 * @param array $instance Widget options for the current instance.
 	 */
-	public function widget( $args, $instance ) {
+	protected function should_hide( $instance ) {
 		$product_id = isset( $instance['product_id'] ) ? $instance['product_id'] : 0;
-		if ( \fastwc_should_hide_pdp_checkout_button( $product_id ) ) {
-			return;
-		}
-
-		$widget_data = array(
-			'template' => 'widgets/fast-product',
-			'args'     => $args,
-			'instance' => $instance,
-		);
-
-		\fastwc_load_template( 'widgets/fast-widget', $widget_data );
+		return \fastwc_should_hide_pdp_checkout_button( $product_id );
 	}
 
 	/**
@@ -52,21 +48,10 @@ class Product extends \WP_Widget {
 	 * @param array $instance Widget options for the current instance.
 	 */
 	public function form( $instance ) {
-		$title       = isset( $instance['title'] ) ? $instance['title'] : '';
-		$product_id  = isset( $instance['product_id'] ) && is_numeric( $instance['product_id'] ) ? $instance['product_id'] : 0;
-		$description = isset( $instance['description'] ) ? $instance['description'] : '';
-		?>
-		<p>
-			<label for="<?php echo esc_attr( $this->get_field_name( 'title' ) ); ?>"><?php esc_html_e( 'Title:', 'fast' ); ?></label>
-			<input
-				class="widefat"
-				id="<?php echo esc_attr( $this->get_field_id( 'title' ) ); ?>"
-				name="<?php echo esc_attr( $this->get_field_name( 'title' ) ); ?>"
-				type="text"
-				value="<?php echo esc_attr( $title ); ?>"
-			/>
-		</p>
+		parent::form( $instance );
 
+		$product_id = isset( $instance['product_id'] ) && is_numeric( $instance['product_id'] ) ? $instance['product_id'] : 0;
+		?>
 		<p>
 			<label for="<?php echo esc_attr( $this->get_field_name( 'product_id' ) ); ?>"><?php esc_html_e( 'Product ID:', 'fast' ); ?></label>
 			<input
@@ -76,10 +61,6 @@ class Product extends \WP_Widget {
 				type="text"
 				value="<?php echo esc_attr( $product_id ); ?>"
 			/>
-		</p>
-		<p>
-			<label for="<?php echo esc_attr( $this->get_field_name( 'description' ) ); ?>"><?php esc_html_e( 'Description:', 'fast' ); ?></label>
-			<?php wp_editor( esc_attr( $description ), $this->get_field_id( 'description' ) ); ?>
 		</p>
 		<?php
 	}
@@ -91,11 +72,9 @@ class Product extends \WP_Widget {
 	 * @param array $old_instance The old settings for the widget.
 	 */
 	public function update( $new_instance, $old_instance ) {
-		$instance = array();
+		$instance = parent::update( $new_instance, $old_instance );
 
-		$instance['title']       = isset( $new_instance['title'] ) ? wp_strip_all_tags( $new_instance['title'] ) : '';
-		$instance['product_id']  = isset( $new_instance['product_id'] ) && is_numeric( $new_instance['product_id'] ) ? $new_instance['product_id'] : 0;
-		$instance['description'] = isset( $new_instance['description'] ) ? $new_instance['description'] : '';
+		$instance['product_id'] = isset( $new_instance['product_id'] ) && is_numeric( $new_instance['product_id'] ) ? $new_instance['product_id'] : 0;
 
 		return $instance;
 	}
